+++ src/Models/InvitationProxy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the InvitationProxy class.
+ *
+ * @copyright   Copyright (c) 2020 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2020-12-18
+ *
+ */
+
+namespace Konekt\User\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class InvitationProxy extends ModelProxy
+{
+}