+++ src/Models/User.php
@@ -14,8 +14,8 @@ namespace Konekt\User\Models;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Authenticatable;
+use Illuminate\Support\Carbon;
 use Konekt\Enum\Eloquent\CastsEnums;
-use Konekt\User\Contracts\Profile;
 use Konekt\User\Contracts\User as UserContract;
 use Konekt\User\Contracts\Profile as ProfileContract;
 use Konekt\User\Events\UserWasActivated;
@@ -24,8 +24,18 @@ use Konekt\User\Events\UserWasDeleted;
 use Konekt\User\Events\UserWasInactivated;
 
 /**
- * User Entity class
- *
+ * @property int $id
+ * @property string $name
+ * @property string $email
+ * @property string $password
+ * @property string|null $remember_token
+ * @property int $login_count
+ * @property UserType $type
+ * @property bool $is_active
+ * @property Carbon|null $last_login_at
+ * @property Carbon $created_at
+ * @property Carbon $updated_at
+ * @property Carbon|null $deleted_at
  */
 class User extends Authenticatable implements UserContract
 {
@@ -59,7 +69,7 @@ class User extends Authenticatable implements UserContract
 
     public function profile()
     {
-        return $this->hasOne(ProfileProxy::modelClass(), 'id', 'id');
+        return $this->hasOne(ProfileProxy::modelClass(), 'user_id', 'id');
     }
 
     public function scopeActive($query)